% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_character}
\alias{enter_character}
\title{Enter a character string}
\usage{
enter_character(prompt = "Enter characters:")
}
\arguments{
\item{prompt}{Message to guide user input}
}
\description{
Enter a character string
}
