% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_classes.R
\name{enter_compliant_value}
\alias{enter_compliant_value}
\title{Enter value with compliance rule.}
\usage{
enter_compliant_value(
  prompt = "Enter compliant value",
  compliance = function(x) {     TRUE },
  fail_message = "Input does not comply."
)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{compliance}{Function which returns a boolean vector.}

\item{fail_message}{Message}
}
\description{
Enter value with compliance rule.
}
