% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_datetime}
\alias{enter_datetime}
\title{Enter a date-time value:}
\usage{
enter_datetime(
  prompt = "Enter date-time in the form of DD/MM/YYYY HH:MM:SS",
  coerce = TRUE
)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to Date-time type?}
}
\description{
Enter a date-time value:
}
