% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_duration}
\alias{enter_duration}
\title{Enter a time duration}
\usage{
enter_duration(prompt = "Enter a duration of time in the form HH:MM:SS")
}
\arguments{
\item{prompt}{Message to guide user input}
}
\description{
Enter a time duration
}
