% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_factor}
\alias{enter_factor}
\title{Enter a factor value:}
\usage{
enter_factor(
  prompt = "Select an option from the list",
  levels,
  print_list = TRUE,
  allow_enter = TRUE,
  coerce = TRUE
)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{levels}{Possible factor levels that can be entered.}

\item{print_list}{Print the options/levels as a list?}

\item{allow_enter}{If there is only a single value, allow the user to press enter to select that value.}

\item{coerce}{Convert to factor?}
}
\description{
Enter a factor value:
}
\examples{
fun <- function() {
  ANSWER <- enter_factor("How satisfied with R are you?",
  levels = c("Extremely", "Quite a lot", "Very","Surprisingly"))
  cat("I knew it.\n")
}
if(interactive()) fun()
}
