% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_integer}
\alias{enter_integer}
\title{Enter an integer}
\usage{
enter_integer(prompt = "Enter an integer value:")
}
\arguments{
\item{prompt}{Message to guide user input}
}
\description{
Enter an integer
}
