% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_multiple}
\alias{enter_multiple}
\title{Enter a value:}
\usage{
enter_multiple(data, auto_values = NULL)
}
\arguments{
\item{data}{Data-set for values list.}

\item{auto_values}{Value here is used instead of prompting for user input.}
}
\description{
Enter a value:
}
