% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_real}
\alias{enter_real}
\title{Enter a floating-point (double precision) number:}
\usage{
enter_real(prompt = "Enter number in a non-ambiguous form:", coerce = TRUE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to Date type?}
}
\description{
Enter a floating-point (double precision) number:
}
\examples{
fun <- function() {
  ANSWER <- enter_real("On a scale of 0 to 1, how satisfied with 1 are you (fractions allowed)?")
  ## a better version would check the answer less cursorily, and
  ## perhaps re-prompt
  if (ANSWER < 1)
    cat("This is impossible.  YOU LIED!\n")
  else
    cat("I knew it.\n")
}
if(interactive()) fun()
}
