% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_value}
\alias{enter_value}
\title{Enter a value:}
\usage{
enter_value(prompt = NULL, template, print_list = TRUE, allow_enter = FALSE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{template}{atomic vector to guess the vector type.}

\item{print_list}{If a factor, print the options/levels as a list?}

\item{allow_enter}{If a factor and there is only a single value, allow the user to press enter to select that value.}
}
\description{
Enter a value:
}
