% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLine2.R
\name{readLine2}
\alias{readLine2}
\title{Version of readlines which works with interactive an command-line modes.}
\usage{
readLine2(message)
}
\arguments{
\item{message}{Message prompt for the user.}
}
\description{
Version of readlines which works with interactive an command-line modes.
}
