% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_boolean}
\alias{enter_boolean}
\title{Enter a TRUE/FALSE value:}
\usage{
enter_boolean(prompt = "Enter TRUE/FALSE or T/F", coerce = TRUE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Coerce to boolean value?}
}
\description{
Enter a TRUE/FALSE value:
}
\examples{
fun <- function() {
  ANSWER <- enter_boolean("True or False: \"I am satisfed with R.\"")
  ## a better version would check the answer less cursorily, and
  ## perhaps re-prompt
  if (!ANSWER)
    cat("This is impossible.  YOU LIED!\n")
  else
    cat("I knew it.\n")
}
if(interactive()) fun()
}
