% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_character}
\alias{enter_character}
\title{Enter a character string}
\usage{
enter_character(prompt = "Enter characters:")
}
\arguments{
\item{prompt}{Message to guide user input}
}
\description{
Enter a character string
}
\examples{
fun <- function() {
  ANSWER <- enter_character("Are you a satisfied R user? ")
  ## a better version would check the answer less cursorily, and
  ## perhaps re-prompt
  if (substr(ANSWER, 1, 1) == "n")
    cat("This is impossible.  YOU LIED!\n")
  else
    cat("I knew it.\n")
}
if(interactive()) fun()
}
