% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_classes.R
\name{enter_compliant_value}
\alias{enter_compliant_value}
\title{Enter value with compliance rule.}
\usage{
enter_compliant_value(
  prompt = "Enter compliant value",
  compliance = function(x) {     TRUE },
  fail_message = "Input does not comply."
)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{compliance}{Function which returns a boolean vector.}

\item{fail_message}{Message}
}
\description{
Enter value with compliance rule.
}
\examples{
fun <- function(){
  enter_compliant_value(prompt = "Enter time in 24 hour format (HH:MM):",
                        compliance = function(x){grepl(pattern = "(0|1|2)\\\\d:[0-5]\\\\d", x)},
                        fail_message = "Input is an invalid time or format")
}
if(interactive()) fun()
}
