% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_date}
\alias{enter_date}
\title{Enter a date}
\usage{
enter_date(prompt = "Enter a date in the form DD/MM/YYYY", coerce = TRUE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to Date type?}
}
\description{
Enter a date
}
