% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_datetime}
\alias{enter_datetime}
\title{Enter a date-time value:}
\usage{
enter_datetime(
  prompt = "Enter date-time in the form of DD/MM/YYYY HH:MM:SS",
  coerce = TRUE
)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to Date-time type?}
}
\description{
Enter a date-time value:
}
\examples{
fun <- function() {
  ANSWER <- enter_datetime("When exactly did you become fully satisfied with R?")
  ## a better version would check the answer less cursorily, and
  ## perhaps re-prompt
  if (ANSWER < "2000-02-29 00:00:00 GMT")
    cat("This is impossible.  YOU LIED!\n")
  else
    cat("I knew it.\n")
}
if(interactive()) fun()
}
