% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_integer}
\alias{enter_integer}
\title{Enter an integer}
\usage{
enter_integer(prompt = "Enter an integer value:", coerce = TRUE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to integer type?}
}
\description{
Enter an integer
}
\examples{
fun <- function() {
  ANSWER <- enter_integer("How many times have you thought about just how satisfying R is? ")
  ## a better version would check the answer less cursorily, and
  ## perhaps re-prompt
  if (ANSWER == 0)
    cat("This is impossible.  YOU LIED!\n")
  else
    cat("I knew it.\n")
}
if(interactive()) fun()
}
