% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter_real}
\alias{enter_real}
\title{Enter a floating-point (double precision) number:}
\usage{
enter_real(prompt = "Enter number in a non-ambiguous form:", coerce = TRUE)
}
\arguments{
\item{prompt}{Message to guide user input}

\item{coerce}{Convert to Date type?}
}
\description{
Enter a floating-point (double precision) number:
}
